# frigappapi/src/main/java/com/frigapp/frigappapi/controller/api/MembersApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.4.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.frigapp.frigappapi.controller.api;

import com.frigapp.frigappapi.model.ErrorResponse;
import com.frigapp.frigappapi.model.FridgesFridgeIdMembersGet200Response;
import com.frigapp.frigappapi.model.FridgesFridgeIdMembersInvitePost201Response;
import com.frigapp.frigappapi.model.InviteMemberRequest;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import jakarta.annotation.Generated;
import jakarta.validation.Valid;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.context.request.NativeWebRequest;

import java.util.Optional;
import java.util.UUID;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-03-26T22:36:13.511413300+01:00[Europe/Paris]", comments = "Generator version: 7.4.0")
@Validated
@Tag(name = "Members", description = "Gestion des collaborateurs d'un frigo")
public interface MembersApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * GET /fridges/{fridgeId}/members : Lister les membres
     *
     * @param fridgeId  (required)
     * @return Liste des membres (status code 200)
     *         or Token JWT absent ou invalide (status code 401)
     *         or Action non autorisée pour ce rôle (status code 403)
     *         or Ressource introuvable (status code 404)
     */
    @Operation(
        operationId = "fridgesFridgeIdMembersGet",
        summary = "Lister les membres",
        tags = { "Members" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Liste des membres", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = FridgesFridgeIdMembersGet200Response.class))
            }),
            @ApiResponse(responseCode = "401", description = "Token JWT absent ou invalide", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "403", description = "Action non autorisée pour ce rôle", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "404", description = "Ressource introuvable", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        },
        security = {
            @SecurityRequirement(name = "BearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/fridges/{fridgeId}/members",
        produces = { "application/json" }
    )
    
    default ResponseEntity<FridgesFridgeIdMembersGet200Response> fridgesFridgeIdMembersGet(
        @Parameter(name = "fridgeId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("fridgeId") UUID fridgeId
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"data\" : [ { \"firstName\" : \"Thomas\", \"role\" : \"collaborator\", \"joinedAt\" : \"2000-01-23T04:56:07.000+00:00\", \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"userId\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"email\" : \"thomas@email.com\" }, { \"firstName\" : \"Thomas\", \"role\" : \"collaborator\", \"joinedAt\" : \"2000-01-23T04:56:07.000+00:00\", \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"userId\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"email\" : \"thomas@email.com\" } ], \"status\" : \"success\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"error\" : { \"code\" : \"FRIDGE_NOT_FOUND\", \"message\" : \"Le frigo demandé n'existe pas.\" }, \"status\" : \"error\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"error\" : { \"code\" : \"FRIDGE_NOT_FOUND\", \"message\" : \"Le frigo demandé n'existe pas.\" }, \"status\" : \"error\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"error\" : { \"code\" : \"FRIDGE_NOT_FOUND\", \"message\" : \"Le frigo demandé n'existe pas.\" }, \"status\" : \"error\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * POST /fridges/{fridgeId}/members/invite : Inviter un membre
     * Invite un utilisateur existant par email. Réservé au propriétaire.
     *
     * @param fridgeId  (required)
     * @param inviteMemberRequest  (required)
     * @return Membre ajouté (status code 201)
     *         or Données invalides (status code 400)
     *         or Token JWT absent ou invalide (status code 401)
     *         or Action non autorisée pour ce rôle (status code 403)
     *         or Utilisateur introuvable (status code 404)
     *         or Déjà membre (status code 409)
     */
    @Operation(
        operationId = "fridgesFridgeIdMembersInvitePost",
        summary = "Inviter un membre",
        description = "Invite un utilisateur existant par email. Réservé au propriétaire.",
        tags = { "Members" },
        responses = {
            @ApiResponse(responseCode = "201", description = "Membre ajouté", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = FridgesFridgeIdMembersInvitePost201Response.class))
            }),
            @ApiResponse(responseCode = "400", description = "Données invalides", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "401", description = "Token JWT absent ou invalide", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "403", description = "Action non autorisée pour ce rôle", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "404", description = "Utilisateur introuvable", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "409", description = "Déjà membre", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        },
        security = {
            @SecurityRequirement(name = "BearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/fridges/{fridgeId}/members/invite",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<FridgesFridgeIdMembersInvitePost201Response> fridgesFridgeIdMembersInvitePost(
        @Parameter(name = "fridgeId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("fridgeId") UUID fridgeId,
        @Parameter(name = "InviteMemberRequest", description = "", required = true) @Valid @RequestBody InviteMemberRequest inviteMemberRequest
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"data\" : { \"firstName\" : \"Thomas\", \"role\" : \"collaborator\", \"joinedAt\" : \"2000-01-23T04:56:07.000+00:00\", \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"userId\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"email\" : \"thomas@email.com\" }, \"status\" : \"success\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"error\" : { \"code\" : \"FRIDGE_NOT_FOUND\", \"message\" : \"Le frigo demandé n'existe pas.\" }, \"status\" : \"error\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"error\" : { \"code\" : \"FRIDGE_NOT_FOUND\", \"message\" : \"Le frigo demandé n'existe pas.\" }, \"status\" : \"error\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"error\" : { \"code\" : \"FRIDGE_NOT_FOUND\", \"message\" : \"Le frigo demandé n'existe pas.\" }, \"status\" : \"error\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"error\" : { \"code\" : \"FRIDGE_NOT_FOUND\", \"message\" : \"Le frigo demandé n'existe pas.\" }, \"status\" : \"error\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"error\" : { \"code\" : \"FRIDGE_NOT_FOUND\", \"message\" : \"Le frigo demandé n'existe pas.\" }, \"status\" : \"error\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * DELETE /fridges/{fridgeId}/members/{userId} : Retirer un membre
     * Retire un collaborateur du frigo. Réservé au propriétaire.
     *
     * @param fridgeId  (required)
     * @param userId  (required)
     * @return Membre retiré (status code 204)
     *         or Token JWT absent ou invalide (status code 401)
     *         or Action non autorisée pour ce rôle (status code 403)
     *         or Ressource introuvable (status code 404)
     */
    @Operation(
        operationId = "fridgesFridgeIdMembersUserIdDelete",
        summary = "Retirer un membre",
        description = "Retire un collaborateur du frigo. Réservé au propriétaire.",
        tags = { "Members" },
        responses = {
            @ApiResponse(responseCode = "204", description = "Membre retiré"),
            @ApiResponse(responseCode = "401", description = "Token JWT absent ou invalide", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "403", description = "Action non autorisée pour ce rôle", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "404", description = "Ressource introuvable", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        },
        security = {
            @SecurityRequirement(name = "BearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/fridges/{fridgeId}/members/{userId}",
        produces = { "application/json" }
    )
    
    default ResponseEntity<Void> fridgesFridgeIdMembersUserIdDelete(
        @Parameter(name = "fridgeId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("fridgeId") UUID fridgeId,
        @Parameter(name = "userId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("userId") UUID userId
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"error\" : { \"code\" : \"FRIDGE_NOT_FOUND\", \"message\" : \"Le frigo demandé n'existe pas.\" }, \"status\" : \"error\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"error\" : { \"code\" : \"FRIDGE_NOT_FOUND\", \"message\" : \"Le frigo demandé n'existe pas.\" }, \"status\" : \"error\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"error\" : { \"code\" : \"FRIDGE_NOT_FOUND\", \"message\" : \"Le frigo demandé n'existe pas.\" }, \"status\" : \"error\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}