# frigappapi/src/main/resources/openapi/openapi_frigapp_v1.0.yaml
openapi: 3.0.3
info:
  title: Frig'App API
  description: |
    API REST de Frig'App — application mobile de gestion de stock de frigo collaboratif.
    
    ## Authentification
    Toutes les routes protégées nécessitent un token JWT passé dans le header `Authorization: Bearer <token>`.
    
    ## Temps réel
    Les modifications sur les items et les membres sont diffusées via WebSocket (STOMP over SockJS) sur `ws://api.frigapp.com/v1/ws`.
    
    ## Conventions
    - Toutes les réponses sont enveloppées dans `{ "status": "success"|"error", "data": ... }`
    - Les identifiants sont des UUID v4
    - Les timestamps sont en ISO 8601 UTC
  version: 1.0.0
  contact:
    name: Équipe Frig'App
servers:
  - url: https://api.frigapp.com/v1
    description: Production
  - url: http://localhost:8080/v1
    description: Développement local (Spring Boot)
...
security: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ── SCHEMAS ─────────────────────────────────────────────────────────────────
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        status:
          type: string
          enum: [success]
          example: success
        data:
          nullable: true
    ErrorResponse:
      type: object
      properties:
        status:
          type: string
          enum: [error]
          example: error
        error:
          type: object
          properties:
            code:
              type: string
              example: FRIDGE_NOT_FOUND
            message:
              type: string
              example: Le frigo demandé n'existe pas.

    # ── USER ──────────────────────────────────────────────────────────────────
...
    MemberSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        firstName:
          type: string
          example: Thomas
        email:
          type: string
          format: email
          example: thomas@email.com
        role:
          type: string
          enum: [owner, collaborator]
          example: collaborator
        joinedAt:
          type: string
          format: date-time
    InviteMemberRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
          example: thomas@email.com

    # ── PRODUCT ───────────────────────────────────────────────────────────────
...
  responses:
    Unauthorized:
      description: Token JWT absent ou invalide
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            status: error
            error:
              code: UNAUTHORIZED
              message: Token JWT absent ou invalide.
    Forbidden:
      description: Action non autorisée pour ce rôle
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            status: error
            error:
              code: FORBIDDEN
              message: Vous n'avez pas les droits pour effectuer cette action.
    NotFound:
      description: Ressource introuvable
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    BadRequest:
      description: Données invalides
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

# ── PATHS ─────────────────────────────────────────────────────────────────────
paths:
...
  /fridges/{fridgeId}/members:
    parameters:
      - name: fridgeId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags: [Members]
      summary: Lister les membres
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Liste des membres
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/MemberSummary'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /fridges/{fridgeId}/members/invite:
    parameters:
      - name: fridgeId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      tags: [Members]
      summary: Inviter un membre
      description: Invite un utilisateur existant par email. Réservé au propriétaire.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InviteMemberRequest'
      responses:
        '201':
          description: Membre ajouté
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/MemberSummary'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: Utilisateur introuvable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                status: error
                error:
                  code: USER_NOT_FOUND
                  message: Aucun compte trouvé avec cet email.
        '409':
          description: Déjà membre
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                status: error
                error:
                  code: ALREADY_MEMBER
                  message: Cet utilisateur est déjà membre de ce frigo.
  /fridges/{fridgeId}/members/{userId}:
    parameters:
      - name: fridgeId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: userId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    delete:
      tags: [Members]
      summary: Retirer un membre
      description: Retire un collaborateur du frigo. Réservé au propriétaire.
      security:
        - BearerAuth: []
      responses:
        '204':
          description: Membre retiré
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ── PRODUCTS ──────────────────────────────────────────────────────────────
...